+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 - [Significant performance improvement in `Natural_Order` and new `Faker`
   methods added to `Standard.Test`][3276]
 - [Implemented `Integer.parse`][3283]
+- [Made `Text.compare_to` correctly handle Unicode normalization][3282]
 
 [debug-shortcuts]:
   https://github.com/enso-org/enso/blob/develop/app/gui/docs/product/shortcuts.md#debug
@@ -74,6 +75,7 @@
 [3276]: https://github.com/enso-org/enso/pull/3276
 [3278]: https://github.com/enso-org/enso/pull/3278
 [3283]: https://github.com/enso-org/enso/pull/3283
+[3282]: https://github.com/enso-org/enso/pull/3282
 
 #### Enso Compiler